CDK-lambda-RDS-CRUD/requirements.txt
aws-cdk-lib
constructs>=10.0.0
psycopg2-binary
boto3